Added the ability to read specificCollider from a library reaction
A specificCollider species could appear in one of the folowing forms:
'A + B (+S) = C + D (+S)`, or `A + B + S = C + D + S`, depending on
the kenetics type of the PDep reaction. RMG first finds a (+S) type
specificCollider, then also looks for a species that appears both in
the reactants and products (regardless of order). If found, verify
that this is the only scpecificCollider (as we currently only support
one specificCollider per reaction), and remove it from the reactants
and products lists. It is saved as a specificCollider attribute of
the Reaction.